libxtreemfs-java: Offset in write/read operations is now long. Makes possible to write/read long sized files.
git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@2986 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ java/servers/src/org/xtreemfs/common/libxtreemfs/FileHandle.java
@@ -39,7 +39,7 @@ public interface FileHandle {
      * 
      * @return Number of bytes read.
      */
-    public abstract int read(UserCredentials userCredentials, byte[] data, int count, int offset)
+    public abstract int read(UserCredentials userCredentials, byte[] data, int count, long offset)
             throws IOException, PosixErrorException, AddressToUUIDNotFoundException;
 
     /**
@@ -68,7 +68,7 @@ public interface FileHandle {
      * 
      * @return Number of bytes written (see @attention above).
      */
-    public abstract int write(UserCredentials userCredentials, byte[] data, int count, int offset)
+    public abstract int write(UserCredentials userCredentials, byte[] data, int count, long offset)
             throws IOException, PosixErrorException, AddressToUUIDNotFoundException;
 
     /**
@@ -101,7 +101,7 @@ public interface FileHandle {
      * @return Number of bytes written (see @attention above).
      */
     public abstract int write(UserCredentials userCredentials, byte[] data, int dataOffset, int count,
-            int offset) throws IOException, PosixErrorException, AddressToUUIDNotFoundException;
+            long offset) throws IOException, PosixErrorException, AddressToUUIDNotFoundException;
 
     /**
      * Flushes pending writes and file size updates (corresponds to a fsync() system call).
